Make the guard a file boundary rather than a thing to remember
Whether a malformed artifact ends the compilation comes down to whether every
read of it happens inside the guard that turns a raise into an answer. Java
cannot say that, and the test that corrupts bytes and watches can only sample
it — the reading that escaped the guard was the lazy half, and a sample that
dies on the eager half stays green over exactly the defect this branch had.
Deepening the sample answers the sample and leaves the rule to be remembered.

So the reading is a file. `SoutherAnnotations` is the whole of it and is
called once, from inside the guard; `ClassFileDeclarations` decides what a
lookup is told and names nothing of the class-file API. A read hoisted out of
the guard has nowhere to go that is not a new import, and a test over this
package's sources says which two files may have one — the writer and the
reader. Nothing else here touches a class file.

Written over the sources for the reason `EveryDiagnosticCodeIsReadableTest`
gives about its own rule: the next one of these is a failing test rather than
a malformed artifact quietly ending a compile again.

The sample goes deeper as well, since it costs nothing: it drains every
annotation value rather than stopping at the names.
